Safari: Rename package.
