+++ learn/src/my/safari/groovy/basic/Closures.groovy
@@ -1,4 +1,4 @@
-package my.safari.groovy
+package my.safari.groovy.basic
 
 /**
  * Created by tdongsi on 4/13/17.
