+++ learn/src/my/safari/groovy/basic/Collection.groovy
@@ -1,4 +1,4 @@
-package my.safari.groovy
+package my.safari.groovy.basic
 
 /**
  * Created by tdongsi on 4/12/17.
